% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcamdebugr.R
\name{calc_derivative}
\alias{calc_derivative}
\title{Calculates the Jacobian matrix}
\usage{
calc_derivative(sd)
}
\arguments{
\item{sd}{(SolutionDebugger) A SolutionDebugger instance}
}
\value{
A matrix where the column and rows are indexed by market names.
}
\description{
Calculates the Jacobian matrix
}
\details{
Calculates the Jacobian matrix from the set of prices currently set in
the solver.  This is a finite difference derivative for each market
represented by column.
}
