% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcamwrapper.R
\name{convert_period_to_year}
\alias{convert_period_to_year}
\title{Convert from a GCAM model period to year}
\usage{
convert_period_to_year(gcam, period)
}
\arguments{
\item{gcam}{(gcam) An initialized GCAM instance}

\item{period}{(integer vector) The model period to convert to year}
}
\value{
(integer vector) The corresponding model year
}
\description{
Convert from a GCAM model period to year
}
