% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcamwrapper.R
\name{convert_year_to_period}
\alias{convert_year_to_period}
\title{Convert from a GCAM model year to model period}
\usage{
convert_year_to_period(gcam, year)
}
\arguments{
\item{gcam}{(gcam) An initialized GCAM instance}

\item{year}{(integer vector) The model year to convert to period}
}
\value{
(integer vector) The corresponding model period
}
\description{
Convert from a GCAM model year to model period
}
