% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcamdebugr.R
\name{evaluate_partial}
\alias{evaluate_partial}
\title{Sets a single price into the model and evaluate it}
\usage{
evaluate_partial(sd, price, index, scaled)
}
\arguments{
\item{sd}{(SolutionDebugger) A SolutionDebugger instance}

\item{price}{(numeric) A price to be set.}

\item{index}{(integer) An R based index (i.e. starts at 1) at which to update the price.}

\item{scaled}{(boolean) If the given price is already scaled or not.  If it is not
it will be before getting set into the solver.}
}
\value{
An array of the F(x) which results from the evaluation.
}
\description{
Sets a single price into the model and evaluate it
}
\details{
Sets a single price into the model and evaluate it and returns the resulting F(x).
This is similar to calling evauluation with reset = True however it can optimize
by knowing only a single market is changing and therefore only calculate the subset
of the model which is affected which can be significantly faster to evaluate.
}
