% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_library.R
\name{find_placeholders}
\alias{find_placeholders}
\title{Find all placeholder in a raw query and return them in a list.}
\usage{
find_placeholders(query_str)
}
\arguments{
\item{query_str}{A raw query string.}
}
\value{
A list of all the placeholders found in the query where the names
are the tags and the values are the types.
}
\description{
Find query placeholders of the form \code{'{tag@type}'} and put them into a
list such as \code{list('tag1' = 'name', 'tag2' = 'year')} for instance.
}
