% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcamwrapper.R
\name{get_current_period}
\alias{get_current_period}
\title{Get the last run GCAM model period}
\usage{
get_current_period(gcam)
}
\arguments{
\item{gcam}{(gcam) An initialized GCAM instance}
}
\value{
(integer) The last period used in `run_to_period` wheter it succeeded or failed
}
\description{
Get the last run GCAM model period
}
