% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcamwrapper.R
\name{get_current_year}
\alias{get_current_year}
\title{Get the last run GCAM model year}
\usage{
get_current_year(gcam)
}
\arguments{
\item{gcam}{(gcam) An initialized GCAM instance}
}
\value{
(integer) The last period used in `run_to_period` wheter it
}
\description{
Get the last run GCAM model year
}
