% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcamwrapper.R
\name{get_data}
\alias{get_data}
\title{Get some arbitrary data out of GCAM}
\usage{
get_data(gcam, query, query_params = list())
}
\arguments{
\item{gcam}{(gcam) An initialized GCAM instance}

\item{query}{(string) A GCAM fusion-ish search path to determine where to get the data.}

\item{query_params}{(list[string] -> array(string)) User options to translate placeholder
expressions in query should it have any.}
}
\value{
A tibble containing the requested data
}
\description{
Get some arbitrary data out of GCAM
}
\details{
Use GCAM Fusion to get some table of data out of GCAM.
}
