% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcamdebugr.R
\name{get_demand}
\alias{get_demand}
\title{Gets the demands}
\usage{
get_demand(sd, scaled)
}
\arguments{
\item{sd}{(SolutionDebugger) A SolutionDebugger instance}

\item{scaled}{(boolean) If the demands should be returned in the scaled form (native
to the solver) or unscaled (native to the model).}
}
\value{
An array of demands indexed by market name.
}
\description{
Gets the demands
}
\details{
Gets the demand of solvable markets as a Series with market names
as the index.
}
