% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcamwrapper.R
\name{get_fx}
\alias{get_fx}
\title{Gets the F(x)}
\usage{
get_fx(sd)
}
\arguments{
\item{sd}{(SolutionDebugger) A SolutionDebugger instance}
}
\value{
An array of F(x) indexed by market name.
}
\description{
Gets the F(x)
}
\details{
Gets the F(X), or supply - demand, of solvable markets as an array
with market names as the index.
}
