% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcamdebugr.R
\name{get_price_scale_factor}
\alias{get_price_scale_factor}
\title{Gets the price scale factor}
\usage{
get_price_scale_factor(sd)
}
\arguments{
\item{sd}{(SolutionDebugger) A SolutionDebugger instance}
}
\value{
An array of price scale factors indexed by market name.
}
\description{
Gets the price scale factor
}
\details{
Gets the price scale factor which attempts to normalize GCAM prices
to ideally be around 1.
}
