% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcamdebugr.R
\name{get_quantity_scale_factor}
\alias{get_quantity_scale_factor}
\title{Gets the quantity scale factor}
\usage{
get_quantity_scale_factor(sd)
}
\arguments{
\item{sd}{(SolutionDebugger) A SolutionDebugger instance}
}
\value{
An array of supply/demand scale factors indexed by market name.
}
\description{
Gets the quantity scale factor
}
\details{
Gets the quantity scale factor which attempts to normalize GCAM supplies
and demands to ideally be around 1.
}
