% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_library.R
\name{get_query}
\alias{get_query}
\title{Look up a query from a YAML file.}
\usage{
get_query(..., query_file = NULL)
}
\arguments{
\item{...}{Strings specifiying the path to the query to lookup.}

\item{query_file}{The path to a YAML file to read if not NULL otherwise
the package query file will be used.}
}
\value{
The raw query from the query file with units set as an attribute
if specified.
}
\description{
The YAML query files will often be organized into nested categories
so users must provide the full path as individual strings such as:
\code{get_query("emissions", "co2_emissions")}.
}
