% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcamwrapper.R
\name{get_scenario_name}
\alias{get_scenario_name}
\title{Get the scenario name}
\usage{
get_scenario_name(gcam)
}
\arguments{
\item{gcam}{(gcam) An initialized GCAM instance}
}
\value{
(string) The name of the scenario
}
\description{
Get the scenario name
}
