% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcamwrapper.R
\name{get_slope}
\alias{get_slope}
\title{Get the "correction" slope}
\usage{
get_slope(sd)
}
\arguments{
\item{sd}{(SolutionDebugger) A SolutionDebugger instance}
}
\value{
An array of the correction slope indexed by market name.
}
\description{
Get the "correction" slope
}
\details{
Get the "correction" slope, which is used by the solver to give
continous behavior when a price for a market falls below the
"lower bound supply" price, or the price below which GCAM would
produce no supply.  The idea being to add a negative supply to
ensure there is always continous behavior and we would ideally match
the slope to what it was just above the "lower bound price" so the
solver can move back into the proper range of prices.
}
