% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcamdebugr.R
\name{get_supply}
\alias{get_supply}
\title{Gets the supplies}
\usage{
get_supply(sd, scaled)
}
\arguments{
\item{sd}{(SolutionDebugger) A SolutionDebugger instance}

\item{scaled}{(boolean) If the supply should be returned in the scaled form (native
to the solver) or unscaled (native to the model).}
}
\value{
An array of supplies indexed by market name.
}
\description{
Gets the supplies
}
\details{
Gets the supply of solvable markets as an array with market names
as the index.
}
