% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_library.R
\docType{data}
\name{PACKAGE_QUERIES}
\alias{PACKAGE_QUERIES}
\title{The parsed query file included with the package}
\format{An object of class \code{list} of length 2.}
\usage{
PACKAGE_QUERIES
}
\description{
The parsed query file included with the package
}
\keyword{datasets}
