% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_library.R
\name{parse_int_query_param}
\alias{parse_int_query_param}
\title{Parse user options for integer operators and generate GCAM Fusion syntax}
\usage{
parse_int_query_param(param_operands, is_get_data)
}
\arguments{
\item{param_operands}{An array containing operators and potentially an operand to be used with that
operator.}

\item{is_get_data}{A boolean if true follows get data symantics and set data if false.}
}
\value{
A GCAM Fusion filter string representing the parameters given.
}
\description{
The currently supported operators are:
`+`: Indicates to read/write to a DataFrame.  Note if `is_get_data` this is always
implied to be set.
`*`: which always matches and so no additional operands are necessary (note if
param_operands is empty this operator is assumed)
The standard comparison operators: `=`, `<`, `<=`, `>`, `>=`. Note if `is_get_data` is true
or the `+` is not set an additional argument must be supplied which is the integer (or a
string that can be converted to integer) to be used as the RHS operand in the comparison.
}
