% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_library.R
\name{parse_str_query_param}
\alias{parse_str_query_param}
\title{Parse user options for string operators and generate GCAM Fusion syntax}
\usage{
parse_str_query_param(param_operands, is_get_data)
}
\arguments{
\item{An}{array containing operators and potentially an operand to be used with that
operator.}
}
\value{
A GCAM Fusion filter string representing the parameters given.
}
\description{
The currently supported operators are:
`+`: Indicates to read/write to a DataFrame.  Note if `is_get_data` this is always
implied to be set.
`*`: which always matches and so no additional operands are necessary (note if
param_operands is empty this operator is assumed)
The operators: `=`, `=~` (regular expression matching). Note if `is_get_data` is true
or the `+` is not set an additional argument must be supplied which is the string to
be used as the RHS operand in the comparison.
}
