% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcamwrapper.R
\name{print_xmldb}
\alias{print_xmldb}
\title{Write the full results to an XML Database}
\usage{
print_xmldb(gcam)
}
\arguments{
\item{gcam}{(gcam) An initialized GCAM instance}
}
\description{
Write the full results to an XML Database
}
\details{
At the moment all of the options governing XML Database output are
locked into the values set in the \code{configuration} as well as the
\code{XMLDBDriver.properties} file.
}
