% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcamwrapper.R
\name{reset_scales}
\alias{reset_scales}
\title{Reset the price and quantity scale factors}
\usage{
reset_scales(sd, price_scale, quantity_scale)
}
\arguments{
\item{sd}{(SolutionDebugger) A SolutionDebugger instance}

\item{price_scale}{A new vector of price scalers to set.}

\item{quantity_scale}{A new vector of quantity scalers to set.}
}
\description{
Reset the price and quantity scale factors
}
\details{
Note, this may invalidate scaled prices / quantities a user may have already stored
and they should re-create them after calling this method
}
