% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcamdebugr.R
\name{runToPeriod}
\alias{runToPeriod}
\title{Run model period}
\usage{
runToPeriod(gcam, period)
}
\arguments{
\item{gcam}{An initialized GCAM instance}

\item{period}{The GCAM model period to run up to}
}
\value{
GCAM instance
}
\description{
Run model period
}
\details{
This will run the given GCAM instance for all periods up
and including the given period
}
