% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcamdebugr.R
\name{set_data}
\alias{set_data}
\title{Set some aribtrary data into GCAM}
\usage{
set_data(gcam, data, path)
}
\arguments{
\item{gcam}{An initialized GCAM instance}

\item{data}{A data.frame with the data to set}

\item{path}{A GCAM fusion-ish search path to determine where to set the data.}
}
\value{
GCAM instance
}
\description{
Set some aribtrary data into GCAM
}
\details{
Use GCAM Fusion to set some table of data into GCAM.
}
