% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcamwrapper.R
\name{set_data_fast}
\alias{set_data_fast}
\title{Set some aribtrary data into GCAM using an optimized routine}
\usage{
set_data_fast(gcam, data, query, query_params = list())
}
\arguments{
\item{gcam}{(gcam) An initialized GCAM instance}

\item{data}{(data.frame) A data.frame with the data to set}

\item{query}{(string) A GCAM fusion-ish search path to determine where to set the data.}

\item{query_params}{(list[string] -> array(string)) User options to translate placeholder}
}
\value{
GCAM instance
}
\description{
Set some aribtrary data into GCAM using an optimized routine
}
\details{
Note this optimized routine is only suitable for exact matching and could be
slower than the general `set_data` in cases where the given data matches a small fraction
of possible values in GCAM.  Under the hood this method essentially uses the given query
to perform a `get_data` call, then does a left join (essentially, using hashes for speed)
on those queried values with the supplied data frame to match in the new values to set
back into GCAM.
}
