% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcamwrapper.R
\name{set_prices}
\alias{set_prices}
\title{Sets an array of prices into the model}
\usage{
set_prices(sd, prices, scaled)
}
\arguments{
\item{sd}{(SolutionDebugger) A SolutionDebugger instance}

\item{prices}{(numeric array) An array of prices to be set.}

\item{scaled}{(boolean) If the given prices are already scaled or not.  If they are not
they will be before getting set into the solver.}
}
\description{
Sets an array of prices into the model
}
\details{
Sets an aray of prices into the model but does not immediately evaluate them.
}
