% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcamwrapper.R
\name{set_scenario_name}
\alias{set_scenario_name}
\title{Set the scenario name}
\usage{
set_scenario_name(gcam, name)
}
\arguments{
\item{gcam}{(gcam) An initialized GCAM instance}

\item{name}{(string) The name of the scenario to set}
}
\description{
Set the scenario name
}
