% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcamdebugr.R
\name{set_slope}
\alias{set_slope}
\title{Set the "correction" slope}
\usage{
set_slope(sd, slope)
}
\arguments{
\item{sd}{(SolutionDebugger) A SolutionDebugger instance}

\item{slope}{(numeric array) An array of negative correction slopes to be set.}
}
\description{
Set the "correction" slope
}
\details{
Set the "correction" slope array into the model.
}
