% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcamdebugr.R
\name{get_data}
\alias{get_data}
\title{Get some arbitrary data out of GCAM}
\usage{
get_data(gcam, path)
}
\arguments{
\item{gcam}{An initialized GCAM instance}

\item{path}{A GCAM fusion-ish search path to determine where to get the data.}
}
\value{
A tibble containing the requested data
}
\description{
Get some arbitrary data out of GCAM
}
\details{
Use GCAM Fusion to get some table of data out of GCAM.
}
