% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcamwrapper.R
\name{print_xmldb}
\alias{print_xmldb}
\title{Write the full results to an XML Database}
\usage{
print_xmldb(gcam, xmldb_location = "")
}
\arguments{
\item{gcam}{(gcam) An initialized GCAM instance}

\item{xmldb_location}{(string) Override the location to write the XMLDB if non-empty}
}
\value{
(string) The XMLDB location, particuarly useful if relying on defaults from the config.
}
\description{
Write the full results to an XML Database
}
\details{
At the moment all of the options governing XML Database output are
locked into the values set in the \code{configuration} as well as the
\code{XMLDBDriver.properties} file.
}
