% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcamwrapper.R
\name{run_to_period}
\alias{run_to_period}
\title{Run model period}
\usage{
run_to_period(gcam, period)
}
\arguments{
\item{gcam}{(gcam) An initialized GCAM instance}

\item{period}{(integer) The GCAM model period to run up to}
}
\value{
GCAM instance
}
\description{
Run model period
}
\details{
This will run the given GCAM instance for all periods up
and including the given period
}
