% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcamwrapper.R
\name{set_data}
\alias{set_data}
\title{Set some aribtrary data into GCAM}
\usage{
set_data(gcam, data, query, query_params = NULL)
}
\arguments{
\item{gcam}{(gcam) An initialized GCAM instance}

\item{data}{(data.frame) A data.frame with the data to set}

\item{query}{(string) A GCAM fusion-ish search path to determine where to set the data.}

\item{query_params}{(list[string] -> array(string)) User options to translate placeholder}
}
\value{
GCAM instance
}
\description{
Set some aribtrary data into GCAM
}
\details{
Use GCAM Fusion to set some table of data into GCAM.
}
